+++ lib/sidekiq/web.rb
@@ -51,6 +51,7 @@ module Sidekiq
       halt 404 unless params[:name]
       @count = (params[:count] || 25).to_i
       @name = params[:name]
+      @queue = Sidekiq::Queue.new(@name)
       (@current_page, @total_size, @messages) = page("queue:#{@name}", params[:page], @count)
       @messages = @messages.map {|msg| Sidekiq.load_json(msg) }
       erb :queue